requirements.txt
pyyaml
pillow
numpy
torch==2.5.1
torchvision==0.20.1
accelerate==1.2.1
diffusers==0.32.1
transformers==4.47.1
peft==0.14.0
huggingface_hub
safetensors
einops==0.8.0
opencv-python==4.10.0.84
